requirements.txt
Flask
flask-cors
gunicorn
gensim==3.*
psycopg2-binary
